refactor: migrates cookie cutter templates to cli
moves the cookie cutter templates for base cmdb generation into the
python executor itself instead of relying on the cloning of another repo

includes the jmespath package in base package dependencies
