+++ hamlet-cli/hamlet/backend/generate/cmdb/account/template/cookiecutter.json
@@ -1,0 +1,7 @@
+{
+    "account_id" : "",
+    "account_name" : "{{cookiecutter.account_id}}",
+    "account_seed" : "{% set seq = ['0','1','2','3','4','5','6','7','8','9','a','b','c','d','e','f','g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v','w','x','y','z'] %}{% for n in range(10) %}{{ seq | random }}{% endfor %}",
+    "provider_type" : [ "aws", "azure" ],
+    "provider_id" : ""
+}
