+++ hamlet-cli/hamlet/backend/generate/cmdb/account/template/{{cookiecutter.account_name}}/infrastructure/operations/shared/credentials.json
@@ -1,0 +1,1 @@
+{}
