+++ hamlet-cli/hamlet/backend/generate/cmdb/product/__init__.py
@@ -1,0 +1,6 @@
+from hamlet.backend.generate import utils
+from . import template as product_cmdb_template
+
+def run(**kwargs):
+
+    utils.cookiecutter(product_cmdb_template, **kwargs)
