+++ hamlet-cli/hamlet/backend/generate/cmdb/product/template/{{cookiecutter.product_name}}/config/settings/shared/settings.json
@@ -1,0 +1,1 @@
+{}
