+++ hamlet-cli/hamlet/backend/generate/cmdb/product/template/{{cookiecutter.product_name}}/config/settings/{{cookiecutter.environment_name}}/{{cookiecutter.segment_name}}/settings.json
@@ -1,0 +1,1 @@
+{}
