+++ hamlet-cli/hamlet/backend/generate/cmdb/product/template/{{cookiecutter.product_name}}/config/solutionsv2/shared/{{cookiecutter.segment_name}}/segment.json
@@ -1,0 +1,6 @@
+{
+  "Segment" : {
+    "Id" : "{{cookiecutter.segment_id}}",
+    "Name" : "{{cookiecutter.segment_name}}"
+  }
+}
