+++ hamlet-cli/hamlet/backend/generate/cmdb/product/template/{{cookiecutter.product_name}}/config/solutionsv2/shared/{{cookiecutter.segment_name}}/solution.json
@@ -1,0 +1,7 @@
+{
+    "Solution" : {
+        "Id" : "solution"
+    },
+    "Tiers" : {
+    }
+}
