+++ hamlet-cli/hamlet/backend/generate/cmdb/product/template/{{cookiecutter.product_name}}/config/solutionsv2/{{cookiecutter.environment_name}}/environment.json
@@ -1,0 +1,6 @@
+{
+  "Environment" : {
+    "Id" : "{{cookiecutter.environment_id}}",
+    "Name" : "{{cookiecutter.environment_name}}"
+  }
+}
