+++ hamlet-cli/hamlet/backend/generate/cmdb/product/template/{{cookiecutter.product_name}}/config/solutionsv2/{{cookiecutter.environment_name}}/{{cookiecutter.segment_name}}/segment.json
@@ -1,0 +1,4 @@
+{
+  "Segment": {
+  }
+}
