+++ hamlet-cli/hamlet/backend/generate/cmdb/product/template/{{cookiecutter.product_name}}/infrastructure/operations/{{cookiecutter.environment_name}}/.gitignore
@@ -1,0 +1,3 @@
+*.plaintext
+*.decrypted
+*.ppk
