+++ hamlet-cli/hamlet/backend/generate/cmdb/product/template/{{cookiecutter.product_name}}/infrastructure/operations/{{cookiecutter.environment_name}}/{{cookiecutter.segment_name}}/credentials.json
@@ -1,0 +1,1 @@
+{}
