+++ hamlet-cli/hamlet/backend/generate/cmdb/tenant/__init__.py
@@ -1,5 +1,6 @@
 from hamlet.backend.generate import utils
-
+from . import template as tenant_cmdb_template
 
 def run(**kwargs):
-    utils.cookiecutter('cmdb', 'tenant', **kwargs)
+
+    utils.cookiecutter(tenant_cmdb_template, **kwargs)
