+++ hamlet-cli/hamlet/backend/generate/cmdb/tenant/template/{{cookiecutter._base_dir}}/{{cookiecutter.tenant_name}}/ipaddressgroups.json
@@ -1,0 +1,4 @@
+{
+    "IPAddressGroups" : {
+    }
+}
