+++ hamlet-cli/hamlet/backend/generate/cmdb/tenant/template/{{cookiecutter._base_dir}}/{{cookiecutter.tenant_name}}/tenant.json
@@ -1,0 +1,16 @@
+{
+    "Tenant" : {
+        "Id" : "{{cookiecutter.tenant_id}}",
+        "Name" : "{{cookiecutter.tenant_name}}"
+    },
+    "Account" : {
+        "Region" : "{{cookiecutter.default_region}}",
+        "Audit" : {
+            "Offline" :  {{cookiecutter.audit_log_offline_days}},
+            "Expiration" :  {{cookiecutter.audit_log_expiry_days}}
+        }
+    },
+    "Product" : {
+        "Region" : "{{cookiecutter.default_region}}"
+    }
+}
