--- hamlet-cli/hamlet/backend/generate/product/__init__.py
@@ -1,2 +1,0 @@
-def run():
-    pass
