--- hamlet-cli/hamlet/backend/generate/product/django/__init__.py
@@ -1,5 +1,0 @@
-from hamlet.backend.generate import utils
-
-
-def run(**kwargs):
-    utils.cookiecutter('products', 'django-cookiecutter', **kwargs)
