+++ hamlet-cli/hamlet/command/generate/__init__.py
@@ -1,6 +1,5 @@
 from hamlet.command import root as cli
 from .cmdb import group as cmdb_generation_group
-from .product import group as product_generation_group
 
 
 @cli.group('generate')
@@ -12,4 +11,3 @@ def group():
 
 
 group.add_command(cmdb_generation_group)
-group.add_command(product_generation_group)
