+++ hamlet-cli/hamlet/command/generate/cmdb.py
@@ -4,6 +4,7 @@ from cookiecutter.exceptions import OutputDirExistsException
 
 from hamlet.backend.generate.cmdb import account as generate_account_backend
 from hamlet.backend.generate.cmdb import tenant as generate_tenant_backend
+from hamlet.backend.generate.cmdb import product as generate_product_backend
 from hamlet.utils import dynamic_option, DynamicCommand
 from hamlet.command.generate import utils
 from hamlet.command.generate import decorators
@@ -13,6 +14,53 @@ from hamlet.command.common.exceptions import CommandError
 @click.group('cmdb')
 def group():  # pragma: no cover
     pass
+
+@group.command('tenant', cls=DynamicCommand)
+@dynamic_option(
+    '--tenant-id',
+    help='The unique Id for the tenant',
+    required=True,
+)
+@dynamic_option(
+    '--tenant-name',
+    help='A more descriptive name for the tenant',
+    default=lambda p: p.tenant_id,
+)
+@dynamic_option(
+    '--default-region',
+    help='The id of the default region to use for deployments in this tenant',
+    default='ap-southeast-2',
+)
+@dynamic_option(
+    '--audit-log-expiry-days',
+    help='How long to keep account audit logs for',
+    type=click.INT,
+    default=2555
+)
+@dynamic_option(
+    '--audit-log-offline-days',
+    help='How log until audit logs are sent to archive storage',
+    type=click.INT,
+    default=90
+)
+@decorators.common_generate_options
+@click.pass_context
+def generate_tenant(
+    ctx,
+    prompt=None,
+    use_default=None,
+    **kwargs
+):
+    """
+    Creates a tenant cmdb.
+    This should be run from the root of an accounts repository.
+    """
+    if not prompt or utils.confirm(kwargs):
+        try:
+            generate_tenant_backend.run(**kwargs)
+
+        except OutputDirExistsException as e:
+            raise CommandError(e)
 
 
 @group.command('account', cls=DynamicCommand)
@@ -56,9 +104,9 @@ def generate_account(
     **kwargs
 ):
     """
-    This template is for the creation of a hamlet account.
-    The account template can be reused to create multiple accounts within a Tenant.
+    Creates an account cmdb within a tenant cmdb.
 
+    The account template can be reused to create multiple accounts within a Tenant.
     This template should be run from a tenant directory
     """
     if not prompt or utils.confirm(kwargs):
@@ -69,49 +117,56 @@ def generate_account(
             raise CommandError(e)
 
 
-@group.command('tenant', cls=DynamicCommand)
+@group.command('product', cls=DynamicCommand)
 @dynamic_option(
-    '--tenant-id',
-    help='The unique Id for the tenant',
+    '--product-id',
+    help='The Id of your product',
     required=True,
 )
 @dynamic_option(
-    '--tenant-name',
-    help='A more descriptive name for the tenant',
-    default=lambda p: p.tenant_id,
+    '--dns-zone',
+    help='A DNS zone name which will act as a base domain for public component hostnames',
+    default='',
 )
 @dynamic_option(
-    '--default-region',
-    help='The id of the default region to use for deployments in this tenant',
-    default='ap-southeast-2',
+    '--product-name',
+    help='A more descriptive name of your product',
+    default=lambda p: p.product_id,
 )
 @dynamic_option(
-    '--audit-log-expiry-days',
-    help='How long to keep account audit logs for',
-    type=click.INT,
-    default=2555
+    '--environment-id',
+    help='The id of your first deployed environment',
+    default='int',
 )
 @dynamic_option(
-    '--audit-log-offline-days',
-    help='How log until audit logs are sent to archive storage',
-    type=click.INT,
-    default=90
+    '--environment-name',
+    help='A more descriptive name of your environment',
+    default='integration',
+)
+@dynamic_option(
+    '--segment-id',
+    help='The id of the first segment in your environments',
+    default='default',
+)
+@dynamic_option(
+    '--segment-name',
+    help='A more descriptive name of your segment',
+    default=lambda p: p.segment_id,
 )
 @decorators.common_generate_options
 @click.pass_context
-def generate_tenant(
+def generate_product(
     ctx,
     prompt=None,
     use_default=None,
     **kwargs
 ):
     """
-    This template is for the creation of a hamlet tenant.
-    This should be run from the root of an accounts repository.
+    Creates a product cmdb.
+    This template should be run from the root of an empty product directory.
     """
     if not prompt or utils.confirm(kwargs):
         try:
-            generate_tenant_backend.run(**kwargs)
-
+            generate_product_backend.run(**kwargs)
         except OutputDirExistsException as e:
             raise CommandError(e)
